info: set the node face
The default is perfectly fine, as long as Emacs thinks the background is
dark.  The function `frame-set-background-mode` does not seem to know
how to pick the background for a vte term.  Perhaps its only method is
through X resources.

The same goes for other libraries whose defaults are still used and
faces depend on the background setting.  For instance,
`org-document-info` and `org-document-title`